refactor(map): MapToolsControl — fold in verified filter
Move the "verified within N years" radio filter into the tools panel as
an optional section, with its loading spinner and active-state dot reused
on the shared tools button. Removes the standalone VerifiedFilterControl
and its registration; verified_filter_change analytics preserved.

e2e Playwright run deferred to end-of-branch (panel spec lands in the
test task). Gate: format, lint (0 warnings), check (0 errors), vitest.
